requirements.txt
fastapi
PyJWT==2.6.0